Reordered the include files and readjusted them so that the user include all necessary headers by including just the Application.h on it's app class. Also fixed the black part of the listbox which wasn't being invalidated.

+++ Windows-Wrapper/Container.cpp
@@ -22,7 +22,7 @@ void Container::Remove(IComponent* component, bool preserveSite)
 		component->m_Site = nullptr;
 	}
 
-	for (int i = 0; i < SiteCount; ++i)
+	for (size_t i = 0; i < SiteCount; ++i)
 	{
 		if (m_Sites[i] == site)
 		{
@@ -61,7 +61,7 @@ void Container::ValidateName(IComponent* component, const std::string& name)
 
 	if (name.empty())
 	{
-		for (int i = 0; i < (std::min)(SiteCount, m_Sites.size()); ++i)
+		for (size_t i = 0; i < (std::min)(SiteCount, m_Sites.size()); ++i)
 		{
 			ISite* s = m_Sites[i];
 			if (s != nullptr &&
